Move tests to subfolder
