+++ src/test/java/info/u_team/gltf_parser/tests/GLTFParserTest.java
@@ -1,10 +1,11 @@
-package info.u_team.gltf_parser;
+package info.u_team.gltf_parser.tests;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import org.junit.jupiter.api.Test;
 
+import info.u_team.gltf_parser.GLTFParser;
 import info.u_team.gltf_parser.generated.gltf.GlTF;
 
 public abstract class GLTFParserTest {
